Two conflicts. spellbook_screen.cpp had upstream fixing the same DBC
warning we fixed on 2026-08-27 under a different name - their tryField is
our fieldOptional, body for body - and their caller additionally reports
when neither column name is present, which ours did not. Resolved to
upstream's side; our call sites move to tryField and fieldOptional is
deleted rather than left as a second name for one lookup.

lua_engine.cpp conflicted structurally, as last time: our widget API is in
lua_widget_api.cpp and upstream's is still here, so their edits inside that
region read as deletions. 3,769 conflict lines carrying 99 lines of change,
five of which are hand-ported into lua_widget_api.cpp - bank slot ids in
SetInventoryItem, the OnHide-hooked toggle guard in Region_Hide, and
EditBox GetMaxLetters/GetMaxBytes.

+++ src/core/entity_spawner_player.cpp
@@ -234,7 +234,13 @@ void EntitySpawner::spawnOnlinePlayer(uint64_t guid,
         if (!sections.exactFace) {
             LOG_WARNING("spawnOnlinePlayer: no DBC face match for face=",
                         static_cast<int>(look.faceId), " skin=", static_cast<int>(look.skinId),
-                        " race=", raceId, " sex=", genderId,
+                        // Cast, because both are uint8_t and the stream writes
+                        // one as a character: race 9 came out as a tab and sex
+                        // 0 as a NUL, which is unreadable and puts a NUL in the
+                        // log file - grep then takes the whole log for binary
+                        // and prints nothing at all for any pattern in it.
+                        " race=", static_cast<int>(raceId),
+                        " sex=", static_cast<int>(genderId),
                         sections.haveFace ? " - using the nearest face instead"
                                           : " - this player will render with no face");
         }
